requirements.txt
flask
requests
python-dotenv
pymongo
xhtml2pdf
authlib
pyotp
openai>=1.12.0
gunicorn
